till 15th june

+++ views/user_dashboard/view_inventory.php
@@ -4,6 +4,7 @@ include("../../includes/header.php");
 include("../../includes/footer.php");
 include("../../config/connection.php");
 require("../../core/functions.php");
+include ('navbar.php');
 ?>
 <h1 class="text-center mt-5">Inventory Management</h1>
 
@@ -60,12 +61,10 @@ require("../../core/functions.php");
                         <div class="mt-3">
                             <table class="table table-striped">
                                 <tr>
-                                    <!-- <th>Item ID</th> -->
                                     <th>SKU</th>
                                     <th>Item Name</th>
                                     <th>Available Stock</th>
                                     <th>Expiry Date</th>
-                                    <!-- <th>Storage Location <br>(Branch)</th> -->
                                     <th>Actions</th>
                                 </tr>
                                 <?php
@@ -75,68 +74,31 @@ require("../../core/functions.php");
                                     while ($row = mysqli_fetch_assoc($allFoodItems)) {
                                 ?>
                                         <tr>
-                                            <!-- <td> -->
                                                 <?php $item_id= $row['item_id'] ?>
-                                            <!-- </td> -->
                                             <td><?php echo $row['sku'] ?></td>
-                                            <?php //echo $row['user_id'] ?>
                                             <td><?php echo $row['item_name'] ?></td>
                                             <td class="text-center"><?php echo $row['quantity'] ?></td>
                                             <td><?php echo getStandardDateFormat($row['expiry_date']) ?></td>
-                                            <!-- <td> -->
-                                                <?php
-                                                // switch($row['storage_location']){
-                                                //     case 'rwp_branch':
-                                                //         echo "Rawawpindi ";
-                                                //         break;
-                                                //     case 'lhr_branch':
-                                                //         echo "Lahore ";
-                                                //         break;
-                                                //     case 'mtn_branch':
-                                                //         echo "Multan ";
-                                                //         break;
-                                                //     case 'khi_branch':
-                                                //         echo "Karachi";
-                                                //         break;
-                                                //     default:
-                                                //         echo '';
-                                                //         break;                                                }
-                                                ?>
-                                            <!-- </td> -->
                                             <td>
-                                                <button class="btn btn-sm btn-primary">
-                                                    <form action="update_item.php?>id=<?php echo $item_id?>">
-                                                    <a href="update_item.php?id=<?php echo $item_id ?>"><i class="bi bi-pencil-square"></i></a>
-                                                    </form>
-                                                </button>
-                                                <button class="btn btn-sm btn-danger del_item">
-                                                    <form action="../../backend/delete_item.php" method="post">
-                                                        <a href="../../backend/delete_item.php?id=<?php echo $item_id ?>"><i class="bi bi-trash"></i></a>
-                                                        </form>
-                                                    </button>
-                                                </button>
+                                                <!-- <button class=""> -->
+                                                    <!-- <form action="update_item.php?>id=<?php //echo $item_id?>"> -->
+                                                    <a href="update_item.php?id=<?php echo $item_id ?>" class="btn btn-sm btn-primary"><i class="bi bi-pencil-square"></i></a>
+                                                    <!-- </form> -->
+                                                <!-- </button> -->
+                                                    <!-- <form action="../../backend/delete_item.php" method="post"> -->
+                                                        <!-- <button> -->
+                                                        <a class="btn btn-sm btn-danger del_item" href="../../backend/delete_item.php?id=<?php echo $item_id ?>" onClick="return confirm('Do you really want to delete this item?')"><i class="bi bi-trash"></i></a>
+                                                        <!-- </form> -->
+                                                        <!-- </button> -->
+                                                <!-- </button> -->
                                             </td>
                                         </tr>
                                 <?php
                                     }
                                 } else {
-                                    // echo "no record";
+                                    
                                 }
                                 ?>
-                                <!-- <tr>
-                                    <td>8152</td>
-                                    <td>Milk Box</td>
-                                    <td class="text-center">40</td>
-                                    <td>22-May-2024</td>
-                                    <td>
-                                        <button class="btn btn-sm btn-primary">
-                                            <i class="bi bi-pencil-square"></i>
-                                        </button>
-                                        <button class="btn btn-sm btn-danger">
-                                            <i class="bi bi-trash"></i>
-                                        </button>
-                                    </td>
-                                </tr> -->
                             </table>
                             <div class="text-end mt-4 mb-1">
                                 <button class="btn btn-danger btn-sm del_all">
@@ -159,13 +121,6 @@ require("../../core/functions.php");
   deleteAllButton.addEventListener('click', function(event) {
     if (!confirm("Are you sure you want to delete all items? This action cannot be undone.")) {
       event.preventDefault();
-    }
-  });
-
-  const deleteItemButton =document.querySelector('.del_item');
-  deleteItemButton.addEventListener('click',function (event){
-    if (!confirm("Are you sure?")){
-        event.preventDefault();
     }
   });
 </script>
